Move v4 jvm code to server
